backport: microblaze: avoid interrupt race conditions
backport of ac58732b8388175dcbe4ef8089d5ae80a8f40fd7

The interrupt acknowledge action have to run after the
registered interrupt handler. So we have a chance to
bear out the corresponding interrupt request in the
corresponding controller hardware.

With this reorderng, we optain a proper interrupt
handling for level triggerd interrupt sources -- for
example the new axi_timer v1.02.a introduced in ISE 13.2

Signed-off-by: Stephan Linz <linz@li-pro.net>